add nimbus feature and add feature attributes where appropriate
