+++ components/nimbus/src/error.rs
@@ -14,6 +14,7 @@ use std::num::{ParseIntError, TryFromIntError};
 pub enum NimbusError {
     #[error("Invalid persisted data")]
     InvalidPersistedData,
+    #[cfg(feature = "nimbus")]
     #[error("Rkv error: {0}")]
     RkvError(#[from] rkv::StoreError),
     #[error("IO error: {0}")]
@@ -50,6 +51,7 @@ pub enum NimbusError {
     DatabaseNotReady,
     #[error("Error parsing a sting into a version {0}")]
     VersionParsingError(String),
+    #[cfg(feature = "nimbus")]
     #[error("Behavior error: {0}")]
     BehaviorError(#[from] BehaviorError),
     #[error("TryFromIntError: {0}")]
@@ -62,6 +64,7 @@ pub enum NimbusError {
     ClientError(#[from] rs_client::ClientError),
 }
 
+#[cfg(feature = "nimbus")]
 #[derive(Debug, thiserror::Error)]
 pub enum BehaviorError {
     #[error("Invalid state: {0}")]
